end of day 3

+++ Day 1 - Band name generator/tip calculator.py
@@ -1,6 +1,6 @@
 print("Welcome to the tip calculator!")
 billTotal = float(input("what was the total bill?"))
-tipPercentage = float(input("How much tip would you like to give? 10,12, or 15?"))
+tipPercentage = int(input("How much tip would you like to give? 10,12, or 15?"))
 tipAmount = billTotal * (tipPercentage/100)
 billTotal += tipAmount
 numPeople = int(input("How many people to split the bill?"))
